Role Moderation (#14)
Resolves #5 

Subcommands `create`, `join`, `leave`, `delete`, `list`. `Delete` will only delete a memberless role. `List` will list members of a role. Users can join a role that they are already in and leave a role that they are already not in with no unerlying effect. 

Added temporary workaround to give all fucntions in `Commands` class access to the message object on which they are acting (`Commands.command_message`). This is protected by a global lock for now. 

Refactored some code to allow for `async` function calls within the `Commands` functions. This required making every `Commands` funciton `async` and adding a few `await` statements.