A few small fixes (#30)
Fixes #29 #26 #29 

* Fix guild verification in !role

Role used the guild before checking if it was None. This caused the role
command to crash memebot when messaging it in a non-guild (i.e. DMs). The none
check on the guild has been moved before guild is used.

* Handle arguments properly in poll

Providing no arguments in poll used to cause a crash, now sends a help 
message. Providing one argument used to add a second thumbsdown argument 
as well, now creates a poll with only one argument.

* Make command detection more strict

The bot used to consider any message starting with a ! to be a command. 
Now, it considers a command to me a message where the first word is a ! 
followed by letters (and not numbers or symbols).

* Change is_command to return bool

* Add commend to is_command to explain regex

* Remove brackets and quotes from !poll help