Command callbacks are called regardless of whether or not command fails (#55)
Added a `status` module to `lib` which lays out status codes for different scenarios. 

`CommandOutput` now accepts an optional `command_status` parameter which expresses the status of the command. Defaults to `status.SUCCESS`.

Command callbacks will now not be run if `Command.exec` returns a `CommandOutput` whose status is not equal to `status.SUCCESS`.