Posts Embeded Quote Tweets and Media (#44)
Co-authored-by: Adam Cooper <adam@adamcooper.sh>

As the first part of our Twitter integration, the bot now scans every incoming message for Twitter URLs, and upon encountering one, performs 2 basic functions:

1. If the bot detects that the posted tweet has multiple images attached, it will react to the message containing the tweet with appropriate emojis displaying the number of images. This is useful for mobile clients for whom embeds don't support multiple images.
2. If the tweet is a quote tweet, it will post a link to the tweet it is quoting. It will perform this process recursively until a depth of 3 quote tweets is reached. This helps to provide context and preserve the comedic timing of quote tweets for all clients, since Discord displays quote tweets as links with no preview.

The Twitter integration currently functions as code in `MemeBot.on_message` which interacts with new class variables in `MemeBot`. We hold on to a `twitter_api` object which allows us to interact with the Twitter API.

Because of this, testing the bot now requires Twitter API credentials which can be obtained here: https://developer.twitter.com/en/portal/dashboard

Credentials should be stored in the root directory in a JSON file called `twitter_api_tokens.json`. They should be stored in the following format:

```json
{
    "consumer_key": string,
    "consumer_secret": string,
    "access_token_key": string,
    "access_token_secret": string,
}
```

It is permissible to include any other information, such as the `bearer_token` in the JSON object, as the required data is loaded on an as-needed basis.