Allow `!role` command to be used with @ mentions (#77)
Previously, the `!role` command would only allow users to input the name of a role, and would error if the user attempted to @ the role. This was inconvenient, because a lot of users use @ mentions for their autocomplete. Now, users can choose to @ the role with which they want to interact, or not! Either will work :)

Also, refactored any uses of `target_name` variables after a `discord.Role` object would have been resolved, so that the name of the role output to the user will always be canonical, and not a long integer ID if it was an @ mention.

Resolves #56.