Make memebot react to itself + fix twitter regex (#76)
The main `on_message()` logic used to `return` immediately when the message was from Memebot itself. Now, it will do all the logic except for executing a command, because we still don't want the bot running commands sent by itself. 

Also, `re.match()` matches only from the beginning of the string, which leads to undesirable behavior in some circumstances (e.g. multiline message with tweet on line 2). Instead, I changed it to use `re.search()`, which scans the whole string for a regex match. 

Fixes #70