Extend `!help` command to accept names of other commands as arguments
The `!help` command will now print longer descriptions and usage instructions for other commands when given their name as an argument. For example:

> !help hello

> A simple ping command. Memebot should respond "Hello!"
>
> **`!hello`**

> !help role list

> List all roles managed by Memebot, or provide the name of a role and list all members of that role.
>
> **`!role list [role]`**

The refactors to the [`Help`](src/commands/help/__init__.py) class were pretty minor, and only include checking for arguments, finding a command that matches the arguments, and printing that command's help text. The behavior for usage with no arguments remains the same.

Several refactors were made to the [`Command`](src/commands/command.py) class to allow it to have more nuanced behavior with regard to helping the user:

1. The abstract method `help_text` has been made non-abstract, and now outputs the message that used to be crafted by `fail`
2. A new abstract method called `long_description` takes the place of `help_text`'s previous purpose.
3. `fail` by default now makes a call to `help_text` and then appends any extra information and sets the `CommandOutput`'s status to `FAIL`

This was done so that we could output the same information previously derived by `fail` without having to regard the command as a failure.