Fix logging in interactive mode
Python is unable to inspect the stack in the interactive repl, so I fixed a bug that would cause the logger to throw every time it attempted to intercept a print statement. The logger will now detect when it is in interactive mode and write the message directly to the true stdout.