Make use of environment variables in Docker builds
Developers can now define environment variables in a .env file to
configure the Docker build. This allows everyone to cater the build
to their different environments/Docker configurations, and to have
separate images for testing and deployment.

Also fixes #94, as one can use the environment variables to define
a specific DNS server.