React to twitter videos with 🎞 if the post is a video
Since Discord removed twitter video embeds, we will now react to video tweets with 🎞 so that people know the tweet is a video and not a photo. We determine it is a video by looking for the string 'video' in its media_url. The API doesn't flag videos differently in any way, so this was the best solution I could come up with.