Post videos in chat when a Tweet has a video (#97)
Since Discord removed twitter video embeds, we will post the raw video link in chat so that people can watch the video without leaving Discord. We determine it is a video by looking for the 'video_info' attribute in its media info dict. The API doesn't flag videos differently in any way, so this was the best solution I could come up with. We then search through all of its variants for the highest bitrate, and post the link to that variant.