Make use of environment variables in Docker builds (#95)
Developers can now define environment variables in a .env file to
configure the Docker build. This allows everyone to cater the build
to their different environments/Docker configurations, and to have
separate images for testing and deployment.

Also fixes #94, as one can use the environment variables to define
a specific DNS server.

Moved a line that copies MemeBot's code to the end of the Dockerfile.
This should make it easier to iterate from build to build, as the
dependencies won't need to be reinstalled every time a source file is
changed.