Use mypy for static type checking (#112)
* Bootstrap test infrastructure

Add a `test` directory with its own `requirements.txt` which is for test dependencies only. Add a `test` target to the `Dockerfile` which installs these dependencies.

* Add mypy cache and `venv/` to the .dockerignore

The mypy cache can get quite large, and it significantly slows down Docker builds if it needs to be copied to the build context.

* Add pyproject.toml

* Add dependency on discord.py-stubs

discord.py does not do a great job at type annotations, so there is a 3rd-party stubs library. The author of this library says that discord.py should have full static type support in 2.0.

See https://github.com/bryanforbes/discord.py-stubs/issues/150#issuecomment-1084611200

* Fix all static type errors in Memebot's source code

* Update README with mypy information