Use GitHub Actions for CI/CD (#120)
Adds a number of workflows to ensure everything we submit to Memebot is easier to maintain. This resolves #32.

### [CodeQL](https://codeql.github.com)
This is GitHub's supported security scanner. It works _very_ well and should catch us while our pants are still firmly belted to our waists.

### [Hadolint](https://github.com/hadolint/hadolint) 
Hadolint is a Dockerfile linter. There are a lot of ways to screw up a Dockerfile, so Hadolint will ensure that we're not doing as many stupid things as we were before #108

### [misspell](https://github.com/marketplace/actions/run-misspell-with-reviewdog)
For catching typos

### [mypy](http://mypy-lang.org)
Since we added the capability to scan our static typing with mypy with #112, we will now enforce it in our CI workflow.

### [Black](https://black.readthedocs.io/en/stable/)
We agreed to use Black as our linter with #18, and then linted the codebase with it in #116, so now we enforce Black's linting rules in CI.

### [yamllint](https://yamllint.readthedocs.io/en/stable/)
I set a pretty lax configuration, but ensures that our YAML files are not a mess.

### [pytest](https://pytest.org/)
Testing will be added with #31, but I set up the action ahead of time. 

### [GitHub Container Registry](https://github.com/features/packages)
When a pull request is merged, GitHub will now build the latest image and publish it on the GitHub Container Registry. This means that there is now an official Memebot image :)

## What is missing
Only half of the desired continuous deployment workflow is in this patch. We build and publish the image, but deploying that image to the server proved to be challenging to do securely with GitHub Actions. I ultimately decided that it would be better to build a separate GitHub webhook consumer on my own for this job. Until then, I'll still have to manually pull down the new image and redploy.