Add workflow for triggering automatic deployment
This workflow runs a simple command under the "Production" environment, which causes GitHub to trigger a deployment event, which is sent to an app I've been developing which can automatically deploy Memebot.

The app takes care of the actual deployment logic, so this workflow is only meant to create the event to trigger the app. Using the app allows me to keep SSH keys on the server, and not have to upload to GitHub or trust that an SSH action that somebody has uploaded will never have malicious code.

This initial commit has a trigger for pull requests just to make sure it works. Before it merges, that trigger will be removed so that it only runs after a successful Docker image build.