Add ability to list roles by user (#128)
By running `!role list <username>`, you can now list the Memebot-managed roles for a given user. The argument to `!role list` can now be either a role or a user, so if there is a user and a role with the same name, the role will take priority when parsing and converting the arguments. 

Co-authored-by: Adam Cooper <adam@adamcooper.sh>