Use `emoji` library for emoji operations (#113)
Emoji are now gathered from the [`emoji`](https://pypi.org/project/emoji/) library, so we don't need to do weird emoji parsing ourselves.

We take on this dependency rather than using `discord.PartialEmoji` because `discord.PartialEmoji` is not a stable API. It is designed to be received by library users, not created. It also does not provide an easy way to insert emoji into text. It is only useful for being passed to other `discord` function.

Resolves #57.