Add unit tests (#161)
This patch fully implements `pytest`-based unit tests for memebot. It
currently tests all the commands plus Twitter integration. Most of the
tests are fairly thorough.

In order for the tests to run correctly, Memebot's code has been moved
from the `src` directory to a root `memebot` package, as is typical for
Python projects. The docker-compose file and README have been updated to
reflect this as well.

Resolves #31