Handle errors from any command type (#308)
The global error handler only properly recognized slash commands. It should now recognize any commands which can be associated with an interaction.

If any command types are added to interactions which don't have a name, then there should be a resulting mypy error to prevent us from shooting ourselves in the foot.